requirements.txt
streamlit==1.41.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.2
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
faiss-cpu==1.9.0.post1
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.16
langchain-community==0.3.16
langchain-core==1.2.28
langchain-google-genai==4.2.1
langchain-huggingface==1.2.1
langchain-ollama==1.1.0
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.45
lxml==6.0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.19.0
networkx==3.6.1
nltk==3.9.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
ollama==0.6.1
orjson==3.11.8
packaging==24.2
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.3.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
pypdf==4.3.1
pypdfium2==5.7.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2026.1.31
python-magic==0.4.27
pytz==2026.1.post1
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.5
referencing==0.37.0
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
streamlit==1.41.1
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.2.2
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
unstructured==0.14.10
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zstandard==0.23.0
